requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.2
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-sslserver==0.22
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
